+++ Userland/Libraries/LibWeb/CMakeLists.txt
@@ -125,6 +125,7 @@ set(SOURCES
     DOM/DOMTokenList.cpp
     DOM/DOMTokenList.idl
     DOM/Document.cpp
+    DOM/DocumentLoading.cpp
     DOM/DocumentFragment.cpp
     DOM/DocumentLoadEventDelayer.cpp
     DOM/DocumentType.cpp
