+++ Userland/Libraries/LibWeb/DOM/DocumentLoading.h
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2020, Andreas Kling <kling@serenityos.org>
+ * Copyright (c) 2023, Aliaksandr Kalenik <kalenik.aliaksandr@gmail.com>
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#pragma once
+
+#include <LibWeb/DOM/Document.h>
+
+namespace Web {
+
+bool parse_document(DOM::Document& document, ByteBuffer const& data);
+
+}
